memory-api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.10
httpx==0.27.0
fastembed==0.4.1
